refactor(#3116): extract vec define write-back emitters to a subsystem module (LOC-budget)
Move __vec_set_elem/__vec_set_len emission out of the codegen barrel
(src/codegen/index.ts) into src/codegen/vec-define-writeback.ts; refresh the
LOC baseline for the intended runtime.ts/object-ops.ts growth. Emission is
byte-identical (defineProperty-free module sha unchanged).


✓
