+++ scripts/loc-budget-baseline.json
@@ -1,7 +1,7 @@
 {
   "generated": "2026-07-09",
   "threshold": 1500,
-  "totalCeiling": 385564,
+  "totalCeiling": 386436,
   "files": {
     "src/codegen-linear/index.ts": 5506,
     "src/codegen-linear/runtime.ts": 3638,
@@ -26,7 +26,7 @@
     "src/codegen/expressions/new-super.ts": 5597,
     "src/codegen/expressions/unary-updates.ts": 2087,
     "src/codegen/generators-native.ts": 4687,
-    "src/codegen/index.ts": 16583,
+    "src/codegen/index.ts": 16598,
     "src/codegen/json-codec-native.ts": 2859,
     "src/codegen/literals.ts": 4228,
     "src/codegen/map-runtime.ts": 2104,
@@ -35,7 +35,7 @@
     "src/codegen/native-strings.ts": 7433,
     "src/codegen/number-format-native.ts": 1712,
     "src/codegen/number-ryu.ts": 1632,
-    "src/codegen/object-ops.ts": 4499,
+    "src/codegen/object-ops.ts": 4592,
     "src/codegen/object-runtime.ts": 9726,
     "src/codegen/parse-number-native.ts": 1838,
     "src/codegen/property-access.ts": 8419,
@@ -57,6 +57,6 @@
     "src/ir/lower.ts": 3532,
     "src/ir/nodes.ts": 2926,
     "src/ir/select.ts": 3223,
-    "src/runtime.ts": 15191
+    "src/runtime.ts": 15489
   }
 }
